refactor: add network-primitives (#1101)
* refactor: add network-primitives

* fix tests

* fix docs

* add re-exports

* Block<T>
