+++ crates/network-primitives/README.md
@@ -1,0 +1,3 @@
+# alloy-network-primitives
+
+Primitive types for Alloy network abstraction.
