+++ crates/network-primitives/src/lib.rs
@@ -1,0 +1,13 @@
+#![doc = include_str!("../README.md")]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/alloy-rs/core/main/assets/alloy.jpg",
+    html_favicon_url = "https://raw.githubusercontent.com/alloy-rs/core/main/assets/favicon.ico"
+)]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
+mod traits;
+pub use traits::{ReceiptResponse, TransactionResponse};
+
+mod block;
+pub use block::{BlockTransactionHashes, BlockTransactions, BlockTransactionsKind};
