+++ crates/rpc-types-eth/Cargo.toml
@@ -26,6 +26,8 @@ alloy-serde.workspace = true
 alloy-consensus = { workspace = true, features = ["std", "serde"] }
 alloy-eips = { workspace = true, features = ["std", "serde"] }
 
+alloy-network-primitives.workspace = true
+
 itertools.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
