+++ crates/rpc-types-eth/src/block.rs
@@ -1,6 +1,7 @@
 //! Block RPC types.
 
 use crate::{ConversionError, Transaction, Withdrawal};
+use alloy_network_primitives::BlockTransactions;
 use alloy_primitives::{Address, BlockHash, Bloom, Bytes, B256, B64, U256};
 use alloy_serde::OtherFields;
 use serde::{ser::Error, Deserialize, Serialize, Serializer};
@@ -204,324 +205,6 @@ impl TryFrom<Header> for alloy_consensus::Header {
     }
 }
 
-/// Block Transactions depending on the boolean attribute of `eth_getBlockBy*`,
-/// or if used by `eth_getUncle*`
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum BlockTransactions<T = Transaction> {
-    /// Full transactions
-    Full(Vec<T>),
-    /// Only hashes
-    Hashes(Vec<B256>),
-    /// Special case for uncle response.
-    Uncle,
-}
-
-impl<T> Default for BlockTransactions<T> {
-    fn default() -> Self {
-        Self::Hashes(Vec::default())
-    }
-}
-
-impl<T> BlockTransactions<T> {
-    /// Check if the enum variant is used for hashes.
-    #[inline]
-    pub const fn is_hashes(&self) -> bool {
-        matches!(self, Self::Hashes(_))
-    }
-
-    /// Fallibly cast to a slice of hashes.
-    pub fn as_hashes(&self) -> Option<&[B256]> {
-        match self {
-            Self::Hashes(hashes) => Some(hashes),
-            _ => None,
-        }
-    }
-
-    /// Returns true if the enum variant is used for full transactions.
-    #[inline]
-    pub const fn is_full(&self) -> bool {
-        matches!(self, Self::Full(_))
-    }
-
-    /// Fallibly cast to a slice of transactions.
-    ///
-    /// Returns `None` if the enum variant is not `Full`.
-    pub fn as_transactions(&self) -> Option<&[T]> {
-        match self {
-            Self::Full(txs) => Some(txs),
-            _ => None,
-        }
-    }
-
-    /// Returns true if the enum variant is used for an uncle response.
-    #[inline]
-    pub const fn is_uncle(&self) -> bool {
-        matches!(self, Self::Uncle)
-    }
-
-    /// Returns an iterator over the transactions (if any). This will be empty
-    /// if the block is an uncle or if the transaction list contains only
-    /// hashes.
-    #[doc(alias = "transactions")]
-    pub fn txns(&self) -> impl Iterator<Item = &T> {
-        self.as_transactions().map(|txs| txs.iter()).unwrap_or_else(|| [].iter())
-    }
-
-    /// Returns an iterator over the transactions (if any). This will be empty if the block is not
-    /// full.
-    pub fn into_transactions(self) -> std::vec::IntoIter<T> {
-        match self {
-            Self::Full(txs) => txs.into_iter(),
-            _ => std::vec::IntoIter::default(),
-        }
-    }
-
-    /// Returns an instance of BlockTransactions with the Uncle special case.
-    #[inline]
-    pub const fn uncle() -> Self {
-        Self::Uncle
-    }
-
-    /// Returns the number of transactions.
-    #[inline]
-    pub fn len(&self) -> usize {
-        match self {
-            Self::Hashes(h) => h.len(),
-            Self::Full(f) => f.len(),
-            Self::Uncle => 0,
-        }
-    }
-
-    /// Whether the block has no transactions.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-}
-
-impl BlockTransactions<Transaction> {
-    /// Converts `self` into `Hashes`.
-    #[inline]
-    pub fn convert_to_hashes(&mut self) {
-        if !self.is_hashes() {
-            *self = Self::Hashes(self.hashes().copied().collect());
-        }
-    }
-
-    /// Converts `self` into `Hashes`.
-    #[inline]
-    pub fn into_hashes(mut self) -> Self {
-        self.convert_to_hashes();
-        self
-    }
-
-    /// Returns an iterator over the transaction hashes.
-    #[deprecated = "use `hashes` instead"]
-    #[inline]
-    pub fn iter(&self) -> BlockTransactionHashes<'_, Transaction> {
-        self.hashes()
-    }
-
-    /// Returns an iterator over references to the transaction hashes.
-    #[inline]
-    pub fn hashes(&self) -> BlockTransactionHashes<'_, Transaction> {
-        BlockTransactionHashes::new(self)
-    }
-
-    /// Returns an iterator over mutable references to the transaction hashes.
-    #[inline]
-    pub fn hashes_mut(&mut self) -> BlockTransactionHashesMut<'_, Transaction> {
-        BlockTransactionHashesMut::new(self)
-    }
-}
-
-impl From<Vec<B256>> for BlockTransactions {
-    fn from(hashes: Vec<B256>) -> Self {
-        Self::Hashes(hashes)
-    }
-}
-
-impl<T> From<Vec<T>> for BlockTransactions<T> {
-    fn from(transactions: Vec<T>) -> Self {
-        Self::Full(transactions)
-    }
-}
-
-/// An iterator over the transaction hashes of a block.
-///
-/// See [`BlockTransactions::hashes`].
-#[derive(Clone, Debug)]
-pub struct BlockTransactionHashes<'a, T>(BlockTransactionHashesInner<'a, T>);
-
-#[derive(Clone, Debug)]
-enum BlockTransactionHashesInner<'a, T = Transaction> {
-    Hashes(std::slice::Iter<'a, B256>),
-    Full(std::slice::Iter<'a, T>),
-    Uncle,
-}
-
-impl<'a, T> BlockTransactionHashes<'a, T> {
-    #[inline]
-    fn new(txs: &'a BlockTransactions<T>) -> Self {
-        Self(match txs {
-            BlockTransactions::Hashes(txs) => BlockTransactionHashesInner::Hashes(txs.iter()),
-            BlockTransactions::Full(txs) => BlockTransactionHashesInner::Full(txs.iter()),
-            BlockTransactions::Uncle => BlockTransactionHashesInner::Uncle,
-        })
-    }
-}
-
-impl<'a> Iterator for BlockTransactionHashes<'a, Transaction> {
-    type Item = &'a B256;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        match &mut self.0 {
-            BlockTransactionHashesInner::Hashes(txs) => txs.next(),
-            BlockTransactionHashesInner::Full(txs) => txs.next().map(|tx| &tx.hash),
-            BlockTransactionHashesInner::Uncle => None,
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match &self.0 {
-            BlockTransactionHashesInner::Full(txs) => txs.size_hint(),
-            BlockTransactionHashesInner::Hashes(txs) => txs.size_hint(),
-            BlockTransactionHashesInner::Uncle => (0, Some(0)),
-        }
-    }
-}
-
-impl ExactSizeIterator for BlockTransactionHashes<'_, Transaction> {
-    #[inline]
-    fn len(&self) -> usize {
-        match &self.0 {
-            BlockTransactionHashesInner::Full(txs) => txs.len(),
-            BlockTransactionHashesInner::Hashes(txs) => txs.len(),
-            BlockTransactionHashesInner::Uncle => 0,
-        }
-    }
-}
-
-impl DoubleEndedIterator for BlockTransactionHashes<'_, Transaction> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        match &mut self.0 {
-            BlockTransactionHashesInner::Full(txs) => txs.next_back().map(|tx| &tx.hash),
-            BlockTransactionHashesInner::Hashes(txs) => txs.next_back(),
-            BlockTransactionHashesInner::Uncle => None,
-        }
-    }
-}
-
-impl<'a> std::iter::FusedIterator for BlockTransactionHashes<'a, Transaction> {}
-
-/// An Iterator over the transaction hashes of a block.
-///
-/// See [`BlockTransactions::hashes_mut`].
-#[derive(Debug)]
-pub struct BlockTransactionHashesMut<'a, T = Transaction>(BlockTransactionHashesInnerMut<'a, T>);
-
-#[derive(Debug)]
-enum BlockTransactionHashesInnerMut<'a, T = Transaction> {
-    Hashes(std::slice::IterMut<'a, B256>),
-    Full(std::slice::IterMut<'a, T>),
-    Uncle,
-}
-
-impl<'a, T> BlockTransactionHashesMut<'a, T> {
-    #[inline]
-    fn new(txs: &'a mut BlockTransactions<T>) -> Self {
-        Self(match txs {
-            BlockTransactions::Hashes(txs) => {
-                BlockTransactionHashesInnerMut::Hashes(txs.iter_mut())
-            }
-            BlockTransactions::Full(txs) => BlockTransactionHashesInnerMut::Full(txs.iter_mut()),
-            BlockTransactions::Uncle => BlockTransactionHashesInnerMut::Uncle,
-        })
-    }
-}
-
-impl<'a> Iterator for BlockTransactionHashesMut<'a, Transaction> {
-    type Item = &'a mut B256;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        match &mut self.0 {
-            BlockTransactionHashesInnerMut::Full(txs) => txs.next().map(|tx| &mut tx.hash),
-            BlockTransactionHashesInnerMut::Hashes(txs) => txs.next(),
-            BlockTransactionHashesInnerMut::Uncle => None,
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match &self.0 {
-            BlockTransactionHashesInnerMut::Full(txs) => txs.size_hint(),
-            BlockTransactionHashesInnerMut::Hashes(txs) => txs.size_hint(),
-            BlockTransactionHashesInnerMut::Uncle => (0, Some(0)),
-        }
-    }
-}
-
-impl ExactSizeIterator for BlockTransactionHashesMut<'_, Transaction> {
-    #[inline]
-    fn len(&self) -> usize {
-        match &self.0 {
-            BlockTransactionHashesInnerMut::Full(txs) => txs.len(),
-            BlockTransactionHashesInnerMut::Hashes(txs) => txs.len(),
-            BlockTransactionHashesInnerMut::Uncle => 0,
-        }
-    }
-}
-
-impl DoubleEndedIterator for BlockTransactionHashesMut<'_, Transaction> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        match &mut self.0 {
-            BlockTransactionHashesInnerMut::Full(txs) => txs.next_back().map(|tx| &mut tx.hash),
-            BlockTransactionHashesInnerMut::Hashes(txs) => txs.next_back(),
-            BlockTransactionHashesInnerMut::Uncle => None,
-        }
-    }
-}
-
-impl<'a> std::iter::FusedIterator for BlockTransactionHashesMut<'a, Transaction> {}
-
-/// Determines how the `transactions` field of [Block] should be filled.
-///
-/// This essentially represents the `full:bool` argument in RPC calls that determine whether the
-/// response should include full transaction objects or just the hashes.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-pub enum BlockTransactionsKind {
-    /// Only include hashes: [BlockTransactions::Hashes]
-    #[default]
-    Hashes,
-    /// Include full transaction objects: [BlockTransactions::Full]
-    Full,
-}
-
-impl From<bool> for BlockTransactionsKind {
-    fn from(is_full: bool) -> Self {
-        if is_full {
-            Self::Full
-        } else {
-            Self::Hashes
-        }
-    }
-}
-
-impl From<BlockTransactionsKind> for bool {
-    fn from(kind: BlockTransactionsKind) -> Self {
-        match kind {
-            BlockTransactionsKind::Full => true,
-            BlockTransactionsKind::Hashes => false,
-        }
-    }
-}
-
 /// Error that can occur when converting other types to blocks
 #[derive(Clone, Copy, Debug, thiserror::Error)]
 pub enum BlockError {
@@ -646,15 +329,6 @@ mod tests {
         let mut bytes = [0u8; 1024];
         rand::thread_rng().fill(bytes.as_mut_slice());
         let _: Header = Header::arbitrary(&mut arbitrary::Unstructured::new(&bytes)).unwrap();
-    }
-
-    #[test]
-    fn test_full_conversion() {
-        let full = true;
-        assert_eq!(BlockTransactionsKind::Full, full.into());
-
-        let full = false;
-        assert_eq!(BlockTransactionsKind::Hashes, full.into());
     }
 
     #[test]
