+++ crates/rpc-types-eth/src/transaction/receipt.rs
@@ -1,6 +1,7 @@
 use crate::Log;
-use alloy_consensus::{AnyReceiptEnvelope, ReceiptEnvelope, TxType};
+use alloy_consensus::{AnyReceiptEnvelope, ReceiptEnvelope, TxReceipt, TxType};
 use alloy_eips::eip7702::SignedAuthorization;
+use alloy_network_primitives::ReceiptResponse;
 use alloy_primitives::{Address, BlockHash, TxHash, B256};
 use alloy_serde::WithOtherFields;
 use serde::{Deserialize, Serialize};
@@ -127,6 +128,24 @@ impl<T> TransactionReceipt<T> {
 /// Alias for a catch-all receipt type.
 #[doc(alias = "AnyTxReceipt")]
 pub type AnyTransactionReceipt = WithOtherFields<TransactionReceipt<AnyReceiptEnvelope<Log>>>;
+
+impl<T: TxReceipt<Log>> ReceiptResponse for TransactionReceipt<T> {
+    fn contract_address(&self) -> Option<alloy_primitives::Address> {
+        self.contract_address
+    }
+
+    fn status(&self) -> bool {
+        self.inner.status()
+    }
+
+    fn block_hash(&self) -> Option<alloy_primitives::BlockHash> {
+        self.block_hash
+    }
+
+    fn block_number(&self) -> Option<u64> {
+        self.block_number
+    }
+}
 
 #[cfg(test)]
 mod test {
